// packages/admin-react/src/pages/finished-store/shot-product/view/data.d.ts
export interface TableListItem {
  id: number;
  avatar: string;
  spu: string;
  sku: string;
  item_id: string;
  product_name: string;
  quantity: number;
  remark: string;
  purchase_order_num: string;
  status: string;
}

export interface TableListItem2 {
  date: string;
  content1: string;
  content2: string;
  content3: string;
}